+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.1.0
+
+### Added
+- Added option to mark/unmark project as focused.
+- Added missing routes.
+
 ## 1.0.1
 
 ### Fixed